+++ tests/frontendIntegration/drf_async/polls/views.py
@@ -148,7 +148,7 @@ def custom_decorator_for_update_jwt():  # type: ignore
 
                     body = request.data  # type: ignore
                     await session_.merge_into_access_token_payload(  # type: ignore
-                        {**clearing, **body},
+                        {**clearing, **body},  # type: ignore
                         {},  # type: ignore
                     )
 
@@ -185,7 +185,7 @@ def custom_decorator_for_update_jwt_with_handle():  # type: ignore
                 body = request.data  # type: ignore
                 await merge_into_access_token_payload(
                     session_.get_handle(),  # type: ignore
-                    {**clearing, **body},
+                    {**clearing, **body},  # type: ignore
                 )
 
                 resp = Response(session_.get_access_token_payload())  # type: ignore